Moved multisig tests into separate file. Added test to check persistence of proposal data throughout upgrade
